Fix: evm wallet connection (#334)
