+++ apps/web/app/(apps)/loans/page.tsx
@@ -1,13 +1,13 @@
+'use client';
+
 import React from 'react';
+import { useXAccounts } from '@sodax/wallet-sdk-react';
 
 export default function LoansPage() {
+  const xAccounts = useXAccounts();
   return (
     <div className="inline-flex flex-col justify-start items-start gap-4">
-      {/* <div className="mix-blend-multiply justify-end">
-        <span className="text-yellow-dark font-bold leading-9 !text-(size:--app-title)">SODAX </span>
-        <span className="text-yellow-dark font-normal font-[shrikhand] leading-9 !text-(size:--app-title)">loans</span>
-      </div>
-      <div className="mix-blend-multiply justify-start text-clay-light font-normal font-['InterRegular'] leading-snug !text-(size:--subtitle)"></div> */}
+      <div>LoansPage {xAccounts['EVM']?.address}</div>
     </div>
   );
 }
