+++ apps/web/app/(apps)/migrate/_components/index.ts
@@ -1,2 +1,5 @@
+export { default as IcxsodaMigration } from './icxsoda-migration';
+export { default as BnusdMigration } from './bnusd-migration';
 export { SuccessDialog } from './success-dialog';
 export { ErrorDialog } from './error-dialog';
+export { default as CurrencyInputPanel, CurrencyInputPanelType } from './currency-input-panel';
