+++ apps/web/app/(apps)/savings/page.tsx
@@ -1,15 +1,13 @@
+'use client';
+
 import React from 'react';
+import { useXAccounts } from '@sodax/wallet-sdk-react';
 
 export default function SavingsPage() {
+  const xAccounts = useXAccounts();
   return (
     <div className="inline-flex flex-col justify-start items-start gap-4">
-      {/* <div className="mix-blend-multiply justify-end">
-        <span className="text-yellow-dark font-bold leading-9 !text-(size:--app-title)">SODAX </span>
-        <span className="text-yellow-dark font-normal font-[shrikhand] leading-9 !text-(size:--app-title)">
-          savings
-        </span>
-      </div>
-      <div className="mix-blend-multiply justify-start text-clay-light font-normal font-['InterRegular'] leading-snug !text-(size:--subtitle)"></div> */}
+      <div>SavingsPage {xAccounts['EVM']?.address}</div>
     </div>
   );
 }
