+++ apps/web/app/test/page.tsx
@@ -1,0 +1,3 @@
+export default function TestPage() {
+  return <div>TestPage</div>;
+}
