+++ apps/web/app/test/swap/page.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { useXAccounts } from '@sodax/wallet-sdk-react';
+
+export default function SwapTestPage() {
+  const xAccounts = useXAccounts();
+  console.log('xAccounts', xAccounts['EVM']);
+  return <div>SwapTestPage {xAccounts['EVM']?.address}</div>;
+}
